arc-tools/discord-bot/requirements.txt
discord.py>=2.3
openai>=1.0
